Adapter: refactor datasource module (renaming, moving)
