+++ adapter/src/datasource/api/DataSourceTriggerEvent.dto.ts
