+++ adapter/src/datasource/api/DatasourceAMQP.dto.ts
