+++ adapter/src/datasource/api/ErrorResponse.dto.ts
